feat(Q_learning/): added Q learning code in python Created Board.py, Global.py, Player.py, main.py. These are in the NewGame folder. Everything is controlled by main.py where it uses all the classes and uses q learning to act as ai.
Within the main.py file there is 2 functions `training()` and `play()`
the to get the dataset to use for naive bayes model, use training function.
If ai_player.pkl does not exist for whatever reason, use the training function
as well. The training function creates the Q table that the Q learning ai uses
and saves it into a .pkl file. It also creates a tttData.csv which contains
every single state plus the move the ai took which can be used to train
ml or dl models. The play function uses the ai_player.pkl file to provide
the Q learning ai with the proper knowledge on how to play the tic tac toe
game. The ai does not require the pretrained pkl file but will perform better
with it. When play function is used, you as the user can play with the q learning ai.