refactor(q_learning_v2.c): Fixed q table not updating and added save to file function I Fixed the q table not updating issues. There was an issue in the `checkState` function where it did not properly checked that the state is in the q table resulting in false return value. There was also an issue in the `defaultQValue` function where the copying of state to the q_table key did not work and would cause segmentation fault due to it being uninitialised. There was another issue at `findQValue` function where it did not check if the state was in Q table resulting in bogus return index.
I also added the `saveQTable` function that serialilses the Q table values and saves
it into a binary file. A `loadQTable` function should be created to deserialise the
values and load it back into the Q table to be used when using it to play against a
human.

feat(q_table.bin): Q Table Binary File
This file contains the serialised data from the Q table. This also contains
the data from the q table after 500 episodes.