refactor (q_learrning.c, q_learning.h): Updated Q-learning algorithm
Inlcuded finding for q value and storing of calculated q value using
learning rate and decay. Amended guiMLmove to only take in char board
as training dataset is in the repo and will be provided when user run
main program in gui.c

refactor(gui.c): Removed training indicator.

Removed training indicator to be passed to guiMLmove to indicate
if the model should be trained or take pre-trained dataset.